[UT] Add JUnit 5 dependencies and refactor type-related tests for improved testing framework compatibility
